backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
PyPDF2==3.0.1
spacy==3.8.0
scikit-learn==1.5.2
numpy==1.26.4
openai==1.51.0
python-jwt==4.0.0
email-validator==2.2.0
pytest==8.3.3
pytest-asyncio==0.24.0

